requirements.txt
blinker==1.9.0
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.0
fonttools==4.57.0
imbalanced-learn==0.13.0
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
